refactor: 코드 수정
