+++ src/page-layout/MyShopLayout/component/MyShop/MyShop.module.scss
@@ -1,5 +1,4 @@
 .container {
-  flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -20,7 +19,7 @@
     width: 96.4rem;
 
     @include tablet {
-      width: 68rem;
+      width: 100%;
     }
 
     @include mobile {
