+++ src/page-layout/MyShopLayout/component/MyShop/ShowMyShop/ShowMyshop.module.scss
@@ -23,6 +23,10 @@
 
     @include tablet {
       width: 100%;
+      min-height: 30.8rem;
+      max-height: 35rem;
+
+      height: 100%;
     }
 
     @include mobile {
