+++ src/page-layout/MyShopLayout/component/RegisteredShop/RegisteredMyShop/RegisteredMyShop.module.scss
@@ -52,19 +52,4 @@
       gap: 0.9rem;
     }
   }
-
-  .size {
-    width: 31.2rem;
-    height: 34.9rem;
-
-    @include tablet {
-      width: 33.2rem;
-      height: 35.9rem;
-    }
-
-    @include mobile {
-      width: 17.1rem;
-      height: 26.1rem;
-    }
-  }
 }
