+++ src/page-layout/MyShopLayout/component/RegisteredShop/RegisteredMyShop/RegisteredMyshop.tsx
@@ -1,27 +1,9 @@
 import classNames from "classnames/bind";
 import styles from "@/page-layout/MyShopLayout/component/RegisteredShop/RegisteredMyShop/RegisteredMyShop.module.scss";
+import { RegisterdShop as RegistseredMyShopProps } from "@/page-layout/MyShopLayout/type";
 import PostContainer from "./PostContainer/PostContainer";
 
 const cn = classNames.bind(styles);
-
-interface RegistseredMyShopProps {
-  lastRef: (node?: Element | null | undefined) => void;
-  myShopData: {
-    item: {
-      id: string;
-      hourlyPay: number;
-      startsAt: string;
-      workhour: number;
-      description: string;
-      closed: boolean;
-      // 추가
-      imageUrl?: string;
-      name?: string;
-      address1?: string;
-      originalHourlyPay?: number;
-    };
-  }[];
-}
 
 export default function RegistseredMyShop({ lastRef, myShopData }: RegistseredMyShopProps) {
   return (
@@ -31,7 +13,7 @@ export default function RegistseredMyShop({ lastRef, myShopData }: RegistseredMy
         <div className={cn("gridContainer")}>
           {myShopData?.map((item) => (
             <div key={item.item.id}>
-              <PostContainer className={cn("size")} myShopData={item} />
+              <PostContainer myShopData={item} />
             </div>
           ))}
         </div>
