+++ src/shared/components/Post/HighPriceRateBadge/HighPriceRateBadge.module.scss
@@ -1,13 +1,16 @@
 .badge {
   display: flex;
-  padding: 1rem 0rem;
+  padding: 1rem 1rem;
   align-items: center;
   border-radius: 2rem;
   background: $color-red-40;
+  white-space: nowrap;
 
   .percentage {
+    display: inline-block;
     color: $color-white;
     text-align: center;
+
     @include sopoqa-14-700;
 
     @include mobile {
@@ -27,6 +30,7 @@
   @include mobile {
     background-color: $color-white;
     padding: 0;
+
     &.red20,
     &.red30,
     &.red40 {
@@ -46,9 +50,11 @@
 .closed,
 .isPast {
   background: $color-gray-20;
+
   .percentage {
     color: $color-white;
   }
+
   @include mobile {
     background-color: $color-white;
     padding: 0;
