+++ src/shared/components/Post/Post.module.scss
@@ -6,8 +6,18 @@
   border-radius: 1.2rem;
   border: 0.1rem solid $color-gray-20;
   background: $color-white;
-  width: 100%;
-  height: 100%;
+  width: 31.2rem;
+  height: 34.9rem;
+
+  @include tablet {
+    width: 33.2rem;
+    height: 35.9rem;
+  }
+
+  @include mobile {
+    width: 17.1rem;
+    height: 26.1rem;
+  }
 
   .clock,
   .location {
@@ -53,7 +63,7 @@
     justify-content: space-between;
     align-items: center;
     color: $color-black;
-    @include sopoqa-24-700;
+    @include sopoqa-20-700;
     @include mobile {
       flex-direction: column;
       align-items: flex-start;
