+++ src/shared/components/Post/Post.tsx
@@ -20,7 +20,6 @@ interface PostProps {
   name?: string;
   address?: string;
   originalHourlyPay?: number;
-  className: string;
 }
 
 export default function Post({
@@ -32,14 +31,13 @@ export default function Post({
   name,
   address,
   originalHourlyPay,
-  className,
 }: PostProps) {
   const currentDate = new Date();
   const endDate = new Date(startsAt);
   const isPast = currentDate > endDate;
 
   return (
-    <div className={cn("postContainer", className, { closed, isPast })}>
+    <div className={cn("postContainer", { closed, isPast })}>
       <div className={cn("imgContainer")}>
         <NoticeMessage isPast={isPast} closed={closed} />
         <Image
