+++ src/shared/components/Post/Skeleton/PostSkeleton.module.scss
@@ -1,9 +1,19 @@
 .Container {
-  width: 100%;
-  height: 100%;
   background-color: #dddbdd;
   border-radius: 4px;
   position: relative;
+  width: 31.2rem;
+  height: 34.9rem;
+
+  @include tablet {
+    width: 33.2rem;
+    height: 35.9rem;
+  }
+
+  @include mobile {
+    width: 17.1rem;
+    height: 26.1rem;
+  }
 }
 
 .skeleton {
