+++ src/shared/components/Post/Skeleton/PostSkeleton.tsx
@@ -3,13 +3,9 @@ import styles from "@/shared/components/Post/Skeleton/PostSkeleton.module.scss";
 
 const cn = classNames.bind(styles);
 
-interface PostSkeletonProps {
-  className: string;
-}
-
-export default function PostSkeleton({ className }: PostSkeletonProps) {
+export default function PostSkeleton() {
   return (
-    <div className={cn("Container", className)}>
+    <div className={cn("Container")}>
       <div className={cn("skeleton")} />
     </div>
   );
