+++ src/shared/components/RootLayout/RootLayout.module.scss
@@ -2,7 +2,9 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-}
-.apply {
   height: 100%;
+
+  .children {
+    flex: 1;
+  }
 }
