+++ src/shared/components/RootLayout/RootLayout.tsx
@@ -1,6 +1,10 @@
 import { ReactNode } from "react";
 import Gnb from "@/common/components/Gnb/Gnb";
 import Footer from "@/common/components/Footer/Footer";
+import styles from "@/shared/components/RootLayout/RootLayout.module.scss";
+import classNames from "classnames/bind";
+
+const cn = classNames.bind(styles);
 
 interface RootLayoutProps {
   children: ReactNode;
@@ -8,9 +12,9 @@ interface RootLayoutProps {
 
 export default function RootLayout({ children }: RootLayoutProps) {
   return (
-    <div>
+    <div className={cn("container")}>
       <Gnb />
-      <main>{children}</main>
+      <main className={cn("children")}>{children}</main>
       <Footer />
     </div>
   );
